Adding Jenkins to docker compose. From now on, it is possible to use jenkins
To use jenkins in this point, it is only needed to run the command
$ docker-compose up

After that, it will start a docker container running jenkins.
You can reach it in your docker machine (localhost if linux) or in your
boot2docker machine if you are using boot2docker.

if you are running in a linux machine access: http://localhost:8080

if you are running in a mac/windows, find out the ip address.

On boot2docker you can run:
$ boot2docker ip

Use the IP given by boot2docker in the port 8080.
Example: http://192.168.59.103:8080