Adding setup for some of the plugins will need in order to have a delivery pipeline
In here we are installing some plugins using a curl command to iteract
with jenkins itself, without any modification.

Note that we are using a new image in the docker compose configuration,
but instead of building it or using it as a continuous container, we are
using it as a sort-living container, used only for configuration. After
the configuration the container will exit with zero and we should have
the jenkins with the new plugins installed.

You can use this approach to create any kind of configuration. It
can be used for example when you are not able to set the configuration
in the image and you don't want to overwrite the entrypoint when
starting the container.