Adding seed job using jenkins-cli.jar
In this step we are introducting another tool that allows us to interact
with jenkins from remote nodes, the jenkins-cli.jar.

We are going to use it in this step to create the job that actually runs
and populate all the other jobs. The new jobs are going to be described
using the plugin job-dls.