Adding first jenkins dsl example
This example pulls a docker image called ruby, that only has ruby
installed. It also git clones a repo of mine called IceCream what has a
Gemfile and a Rake file.

It runs bundle to install all dependencies and then runs the rake
command, what runs a collection of rspec tests. Then it runs all tests
and shares the results (if needed) with our workspace in Jenkins, since
we use a volume that works both ways. The results can be used to set
data to the build execution or to show a report.