Adding first version of the jobs for our dummy pipeline
Here we are adding a lot of empty jobs, just to show how we can build a
pipeline with different jobs and organize it all with groovy code.

This example is not trying to define any standards or commom view of a
pipeline. Each product and team have to decide what are going to be the
steps for its pipelines.

Each job has to have a pipeline definition, in order to allow the
delivey pipeline plugin to organize them in the view. So far, only the
first job, unit_test has a link to another one, code_analysis.

In the next commits we will see the integration happening :)